Integration testing mechanics. (#36)
* First attempt.

Apparently selenium and chromedriver are already installed on the GH
runners. So just need "dash[testing]" as a dev dependency and everything
should just work™️.
https://community.plotly.com/t/dash-integration-testing-with-selenium-and-github-actions/43602

* Fix Windows build.

* Actually make the smoke test be a smoke test for WAZP.

A first integration test that starts a WAZP app server in a thread, and
checks that it can be started error-free.

* Update: the home page title changed on main.

* Adding optional pytables dependency.

* Re-add tests/__init__.py

Both coverage.py and @sfmig seem to need it.